Update examples and examples in README
